# api/docs/openapi/schema/market.yaml
openapi: "3.0.3"
info:
  title: Share Cart API
  version: "1.0"
servers:
  - url: http://localhost:8081

paths:
  /market:
    get:
      summary: Get all markets
      security:
        - Bearer: []
      responses:
        "200":
          description: A list of markets.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/market"
        "default":
          $ref: "http_response.yaml#/components/responses/500InternalServerError"

    post:
      summary: Create a new market
      security:
        - Bearer: []
      requestBody:
        $ref: "#/components/requestBodies/createMarket"
      responses:
        "201":
          description: A new market.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/market"
        "default":
          $ref: "http_response.yaml#/components/responses/500InternalServerError"

  /market/{marketId}:
    get:
      summary: Get a market by ID
      security:
        - Bearer: []
      parameters:
        - name: marketId
          in: path
          required: true
          description: The market ID(uuid).
          schema:
            type: string
          example: "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
      responses:
        "200":
          description: A market.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/market"
        "default":
          $ref: "http_response.yaml#/components/responses/500InternalServerError"

    delete:
      summary: Delete a market by ID
      security:
        - Bearer: []
      parameters:
        - name: marketId
          in: path
          required: true
          description: The market ID(uuid).
          schema:
            type: string
          example: "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
      responses:
        "204":
          description: No content.
        "default":
          $ref: "http_response.yaml#/components/responses/500InternalServerError"

components:
  # RequestBodies
  requestBodies:
    createMarket:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
                example: "market_name"
              description:
                type: string
                example: "market_description"
            required:
              - name
              - description

  # Schemas
  schemas:
    market:
      type: object
      properties:
        id:
          type: string
          description: The market ID(uuid).
          example: "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
        name:
          type: string
          example: "market_name"
        description:
          type: string
          example: "market_description"
        images:
          type: array
          items:
            type: string
          description: The market images uuid.
          example: ["xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"]
      required:
        - id
        - name
        - description
        - images